+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Added google analytics
 -   Added basin delineation workflow
 -   Added fire hydrology workflow
+-   Added trace functionality for query fire perimeters
 
 
 ### Changed
@@ -28,4 +29,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
--   
+-   Layers removed when workflow is finished or exited out of